test(webgraph): isolate the single-thread-pool par_map_fold regression
Co-scheduled with the ordered par_map_fold tests in test_utils_coverage, this test's one-thread Rayon pool perturbs Miri's cooperative scheduler enough to trip experimental Tree Borrows in rayon-core's in_place_scope teardown (reproduced at -Zmiri-seed=6 --test-threads=4; the ordered tests pass once it is isolated). Moving it to its own test binary keeps the deadlock regression under both normal CI and Miri without the cross-test interaction. cargo miri test --workspace now passes.
